refactory: Export component and plugin
