+++ src/index.js
@@ -1,10 +1,3 @@
-import VueDisqus from "./vue-disqus.vue"
-
-export default function install (Vue) {
-  Vue.component('VueDisqus', VueDisqus)
-}
-
-// auto install
-if (typeof window !== 'undefined' && typeof window.Vue !== 'undefined') {
-  window.Vue.use(install)
-}
+import VueDisqus from './VueDisqus'
+export default VueDisqus
+export { default as VueDisqusPlugin } from './plugin'
