+++ src/plugin.js
@@ -1,0 +1,10 @@
+import VueDisqus from "./VueDisqus.vue"
+
+export default function install (Vue) {
+  Vue.component('VueDisqus', VueDisqus)
+}
+
+// auto install
+if (typeof window !== 'undefined' && typeof window.Vue !== 'undefined') {
+  window.Vue.use(install)
+}
